Add some ideas for what we want the new version to look like
Instead of shortcuts being specified manually in the scripts, they
should dynamically support on the fly adding, removing, and changing of
shortcuts.

This would allow for a lot more flexibility in the command's use, giving
us the ability to create a temporary shortcut to a directory that we may
only need for a few minutes or hours, at which point it can be deleted.