qcd: Add `setup` function
Should be run before accessing the database to set up configuration
directories necessary for the command to work.

Right now I'm not using the default directories, instead overriding the
`QCD_DATABASE_FILE` variable on the shell for a custom database
location. Haven't tested this yet, but figured I'd commit it and get
back to it later.