qcd: Add "change" command
Option to change the path referenced by a given shortcut. Uses `sed` to
make the replacement and write out the updated database file. Using the
`-E` flag on `sed` so that the `+` regex operator works. The `-E` flag
works on OS X, but on Linux with GNU sed, it would be the `-r` flag
instead.