Update Bash completion file to work with database-enabled version
Bash completion now works for the new database-backed `qcd`.

Arguments 1 and 2 are completed by searching the `qcd` database.
Argument 3 is completed using default Bash completion by setting
`COMPREPLY` to `()` and adding `-o default` to the `complete` call as
recommended here:

http://stackoverflow.com/questions/12933362/getting-compgen-to-include-slashes-on-directories-when-looking-for-files/19062943#19062943

Copied over the directory variables from the `qcd` script. Not sure if I
should be putting them somewhere else so they're not in two different
places. I probably should. I was also concerned that once the scripts
are sourced, these variables will be in a user's shell environment. That
doesn't feel right, but I'm going to ignore it for now.