qcd: Prefix helper functions with `__qcd_`
Wasn't thinking about the fact that these helper functions get exposed
to the global namespace, which is _such_ a bad idea.

Rename these functions to something specific to the program to make them
harder to randomly come across and to have a better chance of not
overriding something that already exists.