qcd: Allow database file to be symlinked
The `sed -i` calls in the "change" and "remove" commands will overwrite
a symlinked database file with a real file, destroying the symlink.

To prevent this and allow the database file to be symlinked (to, say,
allow it to be tracked in a dotfiles repo), use a redirection technique
as described here:
https://www.cyberciti.biz/faq/howto-prevent-sed-i-from-destroying-symlinks-on-linux-unix/

Using this alternate method, we preserve symlinks, and even create a
backup file as before.